+++ spotify/image.py
@@ -49,7 +49,6 @@ class Image(object):
         self._sp_image = ffi.gc(sp_image, lib.sp_image_release)
 
         self.load_event = threading.Event()
-        self._callback_handles = set()
 
     def __repr__(self):
         return 'Image(%r)' % self.link.uri
@@ -71,10 +70,7 @@ class Image(object):
         # called at all, while callbacks added after load is called
         # immediately.
         handle = ffi.new_handle((callback, self))
-        # TODO Think through the life cycle of the handle object. Can it happen
-        # that we GC the image and handle object, and then later the callback
-        # is called?
-        self._callback_handles.add(handle)
+        spotify._callback_handles.add(handle)
         spotify.Error.maybe_raise(lib.sp_image_add_load_callback(
             self._sp_image, _image_load_callback, handle))
         return handle
@@ -82,7 +78,7 @@ class Image(object):
     @serialized
     def remove_load_callback(self, handle):
         """Remove a callback which was added with :meth:`add_load_callback`."""
-        self._callback_handles.remove(handle)
+        spotify._callback_handles.remove(handle)
         spotify.Error.maybe_raise(lib.sp_image_remove_load_callback(
             self._sp_image, _image_load_callback, handle))
 
@@ -166,7 +162,7 @@ def _image_load_callback(sp_image, handle):
         logger.warning('image_load_callback called without userdata')
         return
     (callback, image) = ffi.from_handle(handle)
-    image._callback_handles.remove(handle)
+    spotify._callback_handles.remove(handle)
     if callback is not None:
         callback(image)