+++ tests/test_artist.py
@@ -245,8 +245,8 @@ class ArtistBrowserTest(unittest.TestCase):
         result = None  # noqa
         tests.gc_collect()
 
-        # FIXME The mock keeps the handle/userdata alive, thus the artist is
-        # kept alive, and this test doesn't test what it is intended to test.
+        # The mock keeps the handle/userdata alive, thus this test doesn't
+        # really test that spotify._callback_handles keeps the handle alive.
         artistbrowse_complete_cb = (
             lib_mock.sp_artistbrowse_create.call_args[0][3])
         userdata = lib_mock.sp_artistbrowse_create.call_args[0][4]